% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lykke_scales.R
\name{lykke_pal}
\alias{lykke_pal}
\title{Fetch palette from lykke_palettes}
\usage{
lykke_pal(palette = "sdu_colors", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'lykke_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colors is reversed.}
}
\description{
Fetch palette from lykke_palettes
}
