% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lykke_colors.R
\docType{data}
\name{lykke_palettes}
\alias{lykke_palettes}
\title{My own color palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
lykke_palettes
}
\description{
My own color palettes
}
\keyword{datasets}
