% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lykke_scales.R
\name{scale_color_lykke}
\alias{scale_color_lykke}
\title{Setup color palette for ggplot2}
\usage{
scale_color_lykke(
  ...,
  palette = "sdu_colors",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE
)
}
\arguments{
\item{...}{additional arguments to pass to scale_color_gradientn}

\item{palette}{Choose from 'lykke_palettes' list}

\item{discrete}{whether to use a discrete color palette}

\item{alpha}{transparency}

\item{reverse}{logical, Reverse the order of the colors?}
}
\description{
Setup color palette for ggplot2
}
